Update barcode-scanner.md
I experienced an issues when no write access rights were present for the attribute- strange, as the microflow did not have entity access rights set. It took me quite a while to figure this out, as I had no logs in sandbox. It took me hours. In other words, it might be handy to indicate this here.

Best regards,

Wouter Penris